+++ coinlib/lib/src/crypto/random.dart
@@ -14,7 +14,7 @@ Uint8List generateRandomBytes(int size) {
   return bytes;
 }
 
-List<T> insertRandom<T>(List<T> list, T element) {
+List<T> insertRandom<T>(Iterable<T> list, T element) {
   final newList = List<T>.from(list);
   newList.insert(Random.secure().nextInt(newList.length+1), element);
   return newList;