+++ coinlib/lib/src/tx/inputs/taproot_input.dart
@@ -25,7 +25,4 @@ abstract class TaprootInput extends WitnessInput {
     details.hashType,
   );
 
-  /// The signed size when SIGHASH_DEFAULT is used for all signatures
-  int? get defaultSignedSize => signedSize;
-
 }